Format demographic report tables

Characteristics can be reordered using the chars argument, but traits within those groups will be ordered as they are in the data. The exception to this is for repeated options (e.g., "Yes" and "No"), which will be ordered in the way they first appear in the data.

demo_table_format(
  data,
  char_var = Characteristic,
  chars = c("Gender", "Race", "Region", "Companion Animals", "Went Fishing",
    "Handled Chickens"),
  char_header_blank = FALSE,
  stripe_fill = "lightblue",
  stripe_color = "white",
  text_color = "darkgray",
  na.rm = T,
  return_html = FALSE,
  include_css = TRUE,
  write = FALSE,
  path = "table.txt",
  ...
)

Arguments

data

Required. A dataframe to be formatted as an html table. This dataframe should have the desired column names, with the exception of the main characteristic variable name, which can be overriden to blank.

char_var

Required. Variable that contains characteristics. Set to Characteristic by default.

chars

A vector of quoted characteristic groups. Table will be ordered in the same order these groups are listed. These should NOT include specific traits (e.g., "Woman"), but instead, only the groups those traits fall into (e.g. "Gender"). These are the rows that will be fill in a different color.

char_header_blank

If TRUE, the column specified in char_var will be set to blank. FALSE by default.

stripe_fill

Color to fill characteristic row. "lightblue" by default.

stripe_color

Color of text in characteristic row. "white" by default.

text_color

Color of text in non-characteristic rows and header. "darkgray" by default.

na.rm

Remove NA values from character columns and replace with blanks. TRUE by default. If FALSE, NA will show up in any cells where it appears in the data you feed into this function.

return_html

If TRUE, returns raw HTML of table. FALSE by default

include_css

If TRUE, returns inline CSS for table formatting. TRUE by default. This is only returned if return_html is also TRUE

write

If TRUE, write results to the file specified in the path argument. FALSE by default.

path

File path to be written to if write is TRUE. "table.txt" in working directory by default.

...

Other arguments

Value

An HTML table or raw HTML

Examples

temp_data <- data.frame(
"Trait" = c("Gender", "Woman or Other", "Man",
    "Income", "Less than $25,000", "$25,000 to $49,999", "$50,000 to $74,999",
    "$75,000 to $99,999", "$100,000 or more",
    "Education", "Less than high school graduate",
    "High school graduate", "Some college or university, no degree",
    "Associate's degree or college diploma", "Bachelor's degree",
    "Master's degree", "Professional or doctoral degree"),
"Percent" = c("", "49%", "51%",
    "", "12%", "20%", "19%",
    "22%", "27%",
    "", "3%",
    "18%", "17%",
    "18%", "32%",
   "10%", "3%"))
demo_table_format(temp_data,
char_var = Trait,
chars = c("Gender", "Income", "Education"))
#> <div id="qvflahbuus" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #qvflahbuus .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: hidden;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: hidden;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #qvflahbuus .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #qvflahbuus .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #qvflahbuus .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #qvflahbuus .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #qvflahbuus .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #qvflahbuus .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #qvflahbuus .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #qvflahbuus .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #qvflahbuus .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #qvflahbuus .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #qvflahbuus .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #qvflahbuus .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #qvflahbuus .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #qvflahbuus .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #qvflahbuus .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #qvflahbuus .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #qvflahbuus .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #qvflahbuus .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #qvflahbuus .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #qvflahbuus .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #qvflahbuus .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #qvflahbuus .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #qvflahbuus .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #qvflahbuus .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #qvflahbuus .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #qvflahbuus .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #qvflahbuus .gt_left {
#>   text-align: left;
#> }
#> 
#> #qvflahbuus .gt_center {
#>   text-align: center;
#> }
#> 
#> #qvflahbuus .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #qvflahbuus .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #qvflahbuus .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #qvflahbuus .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #qvflahbuus .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #qvflahbuus .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 65%;
#> }
#> </style>
#>   <table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" style="background-color: #FFFFFF; color: #333333; font-weight: bold; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Trait</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" style="background-color: #FFFFFF; color: #333333; font-weight: bold; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Percent</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td class="gt_row gt_left" style="background-color: #5FB7E5; color: #FFFFFF; text-align: left; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Gender</td>
#> <td class="gt_row gt_left" style="background-color: #5FB7E5; color: #FFFFFF; text-align: center; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"></td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Woman or Other</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">49%</td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Man</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">51%</td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #5FB7E5; color: #FFFFFF; text-align: left; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Income</td>
#> <td class="gt_row gt_left" style="background-color: #5FB7E5; color: #FFFFFF; text-align: center; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"></td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Less than $25,000</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">12%</td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">$25,000 to $49,999</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">20%</td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">$50,000 to $74,999</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">19%</td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">$75,000 to $99,999</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">22%</td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">$100,000 or more</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">27%</td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #5FB7E5; color: #FFFFFF; text-align: left; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Education</td>
#> <td class="gt_row gt_left" style="background-color: #5FB7E5; color: #FFFFFF; text-align: center; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"></td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Less than high school graduate</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">3%</td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">High school graduate</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">18%</td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Some college or university, no degree</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">17%</td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Associate's degree or college diploma</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">18%</td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Bachelor's degree</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">32%</td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Master's degree</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">10%</td></tr>
#>     <tr><td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: left; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">Professional or doctoral degree</td>
#> <td class="gt_row gt_left" style="background-color: #FFFFFF; color: #333333; text-align: center; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">3%</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>