Paint table elements

This function allows you to change the colors cells and text in the body of the table. Header text should be changed with a table_format function.

Usage

paint_table(
  tbl,
  rows = everything(),
  columns = everything(),
  fill_color = NULL,
  text_color = NULL,
  fill_colour = NULL,
  text_colour = NULL
)

Arguments

tbl

A table made with table_format, demo_table_format, or gt

rows

Index numbers of rows to change. If left blank, every row will be affected. Ranges should be specified using colon notation (e.g., rows = 2:3)

columns

Names or index numbers of rows to change. If left blank, every column will be affected. Ranges should be specified using colon notation (e.g., columns = mpg:cyl or columns = 1:2)

fill_color

color to fill cells with. If blank, the fill won't be changed. You may also use fill_colour.

text_color

color to change text colort o. If blank, the text color won't be changed. You may also use text_colour.

fill_colour

See fill_color.

text_colour

See text_color.

Value

A gt table

Examples

head(mtcars) %>% table_format() %>%
  paint_table(rows = 2:3, columns = 4, fill_color = "pink", text_color = "red")
#> <div id="kzwcasqrxb" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #kzwcasqrxb .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: hidden;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: hidden;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #kzwcasqrxb .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #kzwcasqrxb .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #kzwcasqrxb .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #kzwcasqrxb .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #kzwcasqrxb .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #kzwcasqrxb .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #kzwcasqrxb .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #kzwcasqrxb .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #kzwcasqrxb .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #kzwcasqrxb .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #FFFFFF;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #kzwcasqrxb .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #kzwcasqrxb .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #kzwcasqrxb .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #kzwcasqrxb .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #kzwcasqrxb .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #kzwcasqrxb .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #kzwcasqrxb .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #kzwcasqrxb .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #kzwcasqrxb .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #kzwcasqrxb .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #kzwcasqrxb .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #kzwcasqrxb .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #kzwcasqrxb .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #kzwcasqrxb .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #kzwcasqrxb .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #kzwcasqrxb .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #kzwcasqrxb .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #kzwcasqrxb .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-left: 4px;
#>   padding-right: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #kzwcasqrxb .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #kzwcasqrxb .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #kzwcasqrxb .gt_left {
#>   text-align: left;
#> }
#> 
#> #kzwcasqrxb .gt_center {
#>   text-align: center;
#> }
#> 
#> #kzwcasqrxb .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #kzwcasqrxb .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #kzwcasqrxb .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #kzwcasqrxb .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #kzwcasqrxb .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #kzwcasqrxb .gt_two_val_uncert {
#>   display: inline-block;
#>   line-height: 1em;
#>   text-align: right;
#>   font-size: 60%;
#>   vertical-align: -0.25em;
#>   margin-left: 0.1em;
#> }
#> 
#> #kzwcasqrxb .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#> }
#> 
#> #kzwcasqrxb .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #kzwcasqrxb .gt_slash_mark {
#>   font-size: 0.7em;
#>   line-height: 0.7em;
#>   vertical-align: 0.15em;
#> }
#> 
#> #kzwcasqrxb .gt_fraction_numerator {
#>   font-size: 0.6em;
#>   line-height: 0.6em;
#>   vertical-align: 0.45em;
#> }
#> 
#> #kzwcasqrxb .gt_fraction_denominator {
#>   font-size: 0.6em;
#>   line-height: 0.6em;
#>   vertical-align: -0.05em;
#> }
#> </style>
#>   <table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="background-color: #0092B2; color: #FFFFFF; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">mpg</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="background-color: #0092B2; color: #FFFFFF; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">cyl</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="background-color: #0092B2; color: #FFFFFF; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">disp</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="background-color: #0092B2; color: #FFFFFF; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">hp</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="background-color: #0092B2; color: #FFFFFF; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">drat</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="background-color: #0092B2; color: #FFFFFF; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">wt</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="background-color: #0092B2; color: #FFFFFF; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">qsec</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="background-color: #0092B2; color: #FFFFFF; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">vs</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="background-color: #0092B2; color: #FFFFFF; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">am</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="background-color: #0092B2; color: #FFFFFF; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">gear</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" style="background-color: #0092B2; color: #FFFFFF; font-weight: bold; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;">carb</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>21</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>6</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>160</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>110</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>3.9</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>2.62</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>16.46</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>0</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>1</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>4</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>4</p>
#> </div></td></tr>
#>     <tr><td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>21</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>6</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>160</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFC0CB; color: #FF0000; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>110</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>3.9</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>2.875</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>17.02</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>0</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>1</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>4</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>4</p>
#> </div></td></tr>
#>     <tr><td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>22.8</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>4</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>108</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFC0CB; color: #FF0000; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>93</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>3.85</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>2.32</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>18.61</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>1</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>1</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>4</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>1</p>
#> </div></td></tr>
#>     <tr><td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>21.4</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>6</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>258</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>110</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>3.08</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>3.215</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>19.44</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>1</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>0</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>3</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>1</p>
#> </div></td></tr>
#>     <tr><td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>18.7</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>8</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>360</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>175</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>3.15</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>3.44</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>17.02</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>0</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>0</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>3</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #FFFFFF; color: #333333; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>2</p>
#> </div></td></tr>
#>     <tr><td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>18.1</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>6</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>225</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>105</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>2.76</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>3.46</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>20.22</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>1</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>0</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>3</p>
#> </div></td>
#> <td class="gt_row gt_right" style="background-color: #5FB7E5; color: #FFFFFF; border-left-width: 1px; border-left-style: solid; border-left-color: white; border-right-width: 1px; border-right-style: solid; border-right-color: white; border-top-width: 1px; border-top-style: solid; border-top-color: white; border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: white;"><div class='gt_from_md'><p>1</p>
#> </div></td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>