Fish beliefs data set, abridged

A dataset containing survey results from Canadian participants who were asked to report their beliefs about fishes and if they would pledge to reduce their fish consumption

Format

A data frame with 576 rows and variables:

date

date of survey completion

speed

number of seconds taken to complete survey

gender

gender selected by participant

gender_write_in

additional gender write-in information

gender_group

grouped gender data

race_ethnicity

race selected by participant

race_ethnicity_write_in

additional race write-in information

age

participant age

income

income range selected by participant

education

education selected by participant

latinx

Latinx ethnicity identifier

belief_1

level of agreement: "Fish can learn"

belief_2

level of agreement: "Fish don't mind being in a barren environment"

belief_3

level of agreement: "Fish don't care about being over-crowded"

belief_4

level of agreement: "Fish never find it stressful to be picked up or handled"

ffq_1

Frequency of consumption: Dairy

ffq_2

Frequency of consumption: Chicken

ffq_3

Frequency of consumption: Turkey

ffq_4

Frequency of consumption: Fish and seafood

ffq_5

Frequency of consumption: Pork

ffq_6

Frequency of consumption: Beef

ffq_7

Frequency of consumption: Other meat

ffq_8

Frequency of consumption: Eggs

outcome

pledged to reduce consumption of fish products

Source

https://osf.io/n452j/